fixes for refactoring
