+++ src/payloads/services/index.ts
@@ -4,62 +4,17 @@ import {
   NonEmptyString,
   OrganizationFiscalCode
 } from "italia-ts-commons/lib/strings";
-import { DepartmentName } from "../../../generated/definitions/backend/DepartmentName";
-import { NotificationChannelEnum } from "../../../generated/definitions/backend/NotificationChannel";
 import { OrganizationName } from "../../../generated/definitions/backend/OrganizationName";
 import { PaginatedServiceTupleCollection } from "../../../generated/definitions/backend/PaginatedServiceTupleCollection";
 import { ServiceId } from "../../../generated/definitions/backend/ServiceId";
-import { ServiceMetadata } from "../../../generated/definitions/backend/ServiceMetadata";
-import { ServiceName } from "../../../generated/definitions/backend/ServiceName";
 import { ServicePreference } from "../../../generated/definitions/backend/ServicePreference";
 import { ServicePublic } from "../../../generated/definitions/backend/ServicePublic";
 import { ServiceScopeEnum } from "../../../generated/definitions/backend/ServiceScope";
-import { SpecialServiceCategoryEnum } from "../../../generated/definitions/backend/SpecialServiceCategory";
-import { SpecialServiceMetadata } from "../../../generated/definitions/backend/SpecialServiceMetadata";
-import { StandardServiceCategoryEnum } from "../../../generated/definitions/backend/StandardServiceCategory";
-import { ioDevServerConfig } from "../../config";
+import { getService, getServiceMetadata } from "../../utils/service";
 import { validatePayload } from "../../utils/validator";
-import {
-  frontMatter1CTASiciliaVola,
-  frontMatter2CTA2
-} from "../../utils/variables";
+import { frontMatter2CTA2 } from "../../utils/variables";
 import { IOResponse } from "../response";
 import { specialServices } from "./special";
-
-export const getService = (serviceId: string): ServicePublic => {
-  const service = {
-    department_name: "dev department name" as DepartmentName,
-    organization_fiscal_code: "00514490010" as OrganizationFiscalCode,
-    organization_name: "dev organization name" as OrganizationName,
-    service_id: serviceId,
-    service_name: `${faker.company.bs()}` as ServiceName,
-    available_notification_channels: [
-      NotificationChannelEnum.EMAIL,
-      NotificationChannelEnum.WEBHOOK
-    ],
-    version: 1
-  };
-  return validatePayload(ServicePublic, service);
-};
-
-export const getServiceMetadata = (
-  scope: ServiceScopeEnum
-): ServiceMetadata => {
-  return {
-    description: "demo demo <br/>demo demo <br/>demo demo <br/>demo demo <br/>" as NonEmptyString,
-    scope,
-    address: faker.address.streetAddress() as NonEmptyString,
-    email: faker.internet.email() as NonEmptyString,
-    pec: faker.internet.email() as NonEmptyString,
-    phone: faker.phone.phoneNumber() as NonEmptyString,
-    web_url: faker.internet.url() as NonEmptyString,
-    app_android: faker.internet.url() as NonEmptyString,
-    app_ios: faker.internet.url() as NonEmptyString,
-    tos_url: faker.internet.url() as NonEmptyString,
-    privacy_url: faker.internet.url() as NonEmptyString,
-    category: StandardServiceCategoryEnum.STANDARD
-  };
-};
 
 export const getServices = (
   national: number,
